backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
openai>=1.0.0
python-magic-bin==0.4.14
Pillow>=10.0.0
openpyxl>=3.0.0
PyPDF2>=3.0.0
